--- .idea/libraries/DiscordIPC.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="DiscordIPC">
-    <CLASSES>
-      <root url="jar://H:/Pobrane/DiscordIPC.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
